pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "duckduckgo-search>=8.1.1",
    "google-adk==1.5.0",
    "pyarxiv>=1.0.3.1",
    "requests>=2.32.4",
    "wikipedia>=1.4.0",
]
